:recycle: api: move SCO orga certificcation result route to src
