+++ api/lib/application/organizations/organization-controller.js
@@ -110,28 +110,6 @@ const getOrganizationMemberIdentities = async function (
   return dependencies.organizationMemberIdentitySerializer.serialize(members);
 };
 
-const downloadCertificationResults = async function (
-  request,
-  h,
-  dependencies = { getDivisionCertificationResultsCsv },
-) {
-  const organizationId = request.params.id;
-  const { division } = request.query;
-
-  const certificationResults = await usecases.getScoCertificationResultsByDivision({ organizationId, division });
-
-  const csvResult = await dependencies.getDivisionCertificationResultsCsv({
-    division,
-    certificationResults,
-    i18n: request.i18n,
-  });
-
-  return h
-    .response(csvResult.content)
-    .header('Content-Type', 'text/csv;charset=utf-8')
-    .header('Content-Disposition', `attachment; filename="${csvResult.filename}"`);
-};
-
 const attachChildOrganization = async function (request, h) {
   const { childOrganizationId } = request.payload;
   const { organizationId: parentOrganizationId } = request.params;
@@ -261,7 +239,6 @@ const organizationController = {
   cancelOrganizationInvitation,
   create,
   createInBatch,
-  downloadCertificationResults,
   findChildrenOrganizationsForAdmin,
   findPaginatedCampaignManagements,
   findPaginatedFilteredMemberships,
