+++ api/lib/infrastructure/repositories/sco-certification-candidate-repository.js
@@ -21,27 +21,7 @@ const addNonEnrolledCandidatesToSession = async function ({ sessionId, scoCertif
   await knex.batchInsert('certification-candidates', candidatesToBeEnrolledDTOs);
 };
 
-const findIdsByOrganizationIdAndDivision = async function ({ organizationId, division }) {
-  const rows = await knex
-    .select(['certification-candidates.id'])
-    .from('certification-candidates')
-    .join(
-      'view-active-organization-learners',
-      'view-active-organization-learners.id',
-      'certification-candidates.organizationLearnerId',
-    )
-    .where({
-      'view-active-organization-learners.organizationId': organizationId,
-      'view-active-organization-learners.isDisabled': false,
-    })
-    .whereRaw('LOWER("view-active-organization-learners"."division") = ?', division.toLowerCase())
-    .orderBy('certification-candidates.lastName', 'ASC')
-    .orderBy('certification-candidates.firstName', 'ASC');
-
-  return rows.map((row) => row.id);
-};
-
-export { addNonEnrolledCandidatesToSession, findIdsByOrganizationIdAndDivision };
+export { addNonEnrolledCandidatesToSession };
 
 function _scoCandidateToDTOForSession(sessionId) {
   return (scoCandidate) => {
