+++ api/src/certification/course/application/organization-route.js
@@ -1,0 +1,39 @@
+import Joi from 'joi';
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
+import { organizationController } from './organization-controller.js';
+
+const register = async function (server) {
+  server.route([
+    {
+      method: 'GET',
+      path: '/api/organizations/{id}/certification-results',
+      config: {
+        pre: [
+          {
+            method: securityPreHandlers.checkUserIsAdminInSCOOrganizationManagingStudents,
+            assign: 'belongsToOrganizationManagingStudents',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.organizationId,
+          }),
+          query: Joi.object({
+            division: Joi.string().optional(),
+            lang: Joi.string().optional().valid('fr', 'en'),
+          }),
+        },
+        handler: organizationController.downloadCertificationResults,
+        tags: ['api', 'organizations'],
+        notes: [
+          'Cette route est restreinte aux utilisateurs authentifiés',
+          "Elle retourne les certifications liées à l'organisation sous forme de fichier CSV.",
+        ],
+      },
+    },
+  ]);
+};
+
+const name = 'organization';
+export { register, name };
