+++ api/src/certification/course/domain/usecases/index.js
@@ -8,6 +8,8 @@ import { injectDependencies } from '../../../../shared/infrastructure/utils/depe
 import * as assessmentResultRepository from '../../../../shared/infrastructure/repositories/assessment-result-repository.js';
 import * as competenceMarkRepository from '../../../../../lib/infrastructure/repositories/competence-mark-repository.js';
 import * as courseAssessmentResultRepository from '../../infrastructure/repositories/course-assessment-result-repository.js';
+import * as certificationResultRepository from '../../../../../lib/infrastructure/repositories/certification-result-repository.js';
+import * as scoCertificationCandidateRepository from '../../infrastructure/repositories/sco-certification-candidate-repository.js';
 
 /**
  * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
@@ -15,18 +17,21 @@ import * as courseAssessmentResultRepository from '../../infrastructure/reposito
  * @typedef {assessmentResultRepository} AssessmentResultRepository
  * @typedef {courseAssessmentResultRepository} CourseAssessmentResultRepository
  * @typedef {competenceMarkRepository} CompetenceMarkRepository
- *
+ * @typedef {certificationResultRepository} CertificationResultRepository
+ * @typedef {scoCertificationCandidateRepository} ScoCertificationCandidateRepository *
  **/
 const dependencies = {
   assessmentResultRepository,
   competenceMarkRepository,
   courseAssessmentResultRepository,
+  certificationResultRepository,
+  scoCertificationCandidateRepository,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
 
 /**
- * Note : current ignoredFileNames are injected in * {@link file://./../../../../shared/domain/usecases/index.js}
+ * Note : current ignoredFileNames are injected in * {@link file://./../../../shared/domain/usecases/index.js}
  * This is in progress, because they should be injected in this file and not by shared sub-domain
  * The only remaining file ignored should be index.js
  */
