+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -1279,50 +1279,6 @@ describe('Acceptance | Application | organization-controller', function () {
     });
   });
 
-  describe('GET /api/organizations/{id}/certification-results', function () {
-    it('should return HTTP status 200', async function () {
-      // given
-      const user = databaseBuilder.factory.buildUser.withRawPassword();
-
-      const organization = databaseBuilder.factory.buildOrganization({ type: 'SCO', isManagingStudents: true });
-      databaseBuilder.factory.buildMembership({
-        organizationId: organization.id,
-        userId: user.id,
-        organizationRole: Membership.roles.ADMIN,
-      });
-
-      const organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId: organization.id,
-        division: 'aDivision',
-      });
-      const candidate = databaseBuilder.factory.buildCertificationCandidate({
-        organizationLearnerId: organizationLearner.id,
-      });
-      const certificationCourse = databaseBuilder.factory.buildCertificationCourse({
-        userId: candidate.userId,
-        sessionId: candidate.sessionId,
-        isPublished: true,
-      });
-
-      databaseBuilder.factory.buildAssessmentResult.last({
-        certificationCourseId: certificationCourse.id,
-      });
-      await databaseBuilder.commit();
-
-      const options = {
-        method: 'GET',
-        url: `/api/organizations/${organization.id}/certification-results?division=aDivision`,
-        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
-      };
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
-
   describe('GET /api/organizations/{organizationId}/divisions', function () {
     it('should return the divisions', async function () {
       // given
