+++ api/tests/certification/course/unit/domain/usecases/get-sco-certification-results-by-division_test.js
@@ -1,8 +1,8 @@
-import { expect, sinon, domainBuilder, catchErr } from '../../../test-helper.js';
-import { getScoCertificationResultsByDivision } from '../../../../lib/domain/usecases/get-sco-certification-results-by-division.js';
-import { NoCertificationResultForDivision } from '../../../../lib/domain/errors.js';
+import { expect, sinon, domainBuilder, catchErr } from '../../../../../test-helper.js';
+import { getScoCertificationResultsByDivision } from '../../../../../../src/certification/course/domain/usecases/get-sco-certification-results-by-division.js';
+import { NoCertificationResultForDivision } from '../../../../../../lib/domain/errors.js';
 
-describe('Unit | UseCase | get-sco-certification-results-by-division', function () {
+describe('Certification | Course | Unit | UseCase | get-sco-certification-results-by-division', function () {
   const scoCertificationCandidateRepository = { findIdsByOrganizationIdAndDivision: null };
   const certificationResultRepository = { findByCertificationCandidateIds: null };
   const dependencies = {
